# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Tasing API",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Tasing API Server"
        }
    ],
    "paths": {
        "/api/password/email": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Send a reset link to the given user",
                "operationId": "sendResetLinkEmail",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Password reset email sended.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "type": "string"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "The given data was invalid."
                    }
                }
            }
        },
...
        "/api/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Handle a registration request for the application",
                "operationId": "register",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "lastname",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password_confirmation",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User registered successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "properties": {
                                                "user": {
                                                    "$ref": "#/components/schemas/User"
                                                },
                                                "access_token": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "The given data was invalid."
                    }
                }
            }
        },
        "/api/password/reset": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Reset the given user's password",
                "operationId": "reset",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password_confirmation",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "token",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Password reseted.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "type": "string"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "The given data was invalid."
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "User": {
                "properties": {
                    "id": {
                        "description": "id",
                        "type": "integer",
                        "format": "int32"
                    },
                    "name": {
                        "description": "name",
                        "type": "string"
                    },
                    "lastname": {
                        "description": "lastname",
                        "type": "string"
                    },
                    "email": {
                        "description": "email",
                        "type": "string"
                    },
                    "phone_number1": {
                        "description": "phone_number1",
                        "type": "string"
                    },
                    "phone_number2": {
                        "description": "phone_number2",
                        "type": "string"
                    },
                    "country_code": {
                        "description": "country_code",
                        "type": "string"
                    },
                    "address_line1": {
                        "description": "address_line1",
                        "type": "string"
                    },
                    "address_line2": {
                        "description": "address_line2",
                        "type": "string"
                    },
                    "city": {
                        "description": "city",
                        "type": "string"
                    },
                    "region": {
                        "description": "region",
                        "type": "string"
                    },
                    "zipcode": {
                        "description": "zipcode",
                        "type": "string"
                    },
                    "company_name": {
                        "description": "company_name",
                        "type": "string"
                    },
                    "company_number": {
                        "description": "company_number",
                        "type": "string"
                    },
                    "email_verified_at": {
                        "description": "email_verified_at",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        }
    }
}